Make ReservationsTable a container
Makes it a lot easier to add new features to
ReservationsTable now that props don't need to be passed
through UserReservationsPage container. Also makes it super
easy to port the ReservationsTable to other places in the
site if need ever rises.
Refs #89.
